ai-coding-assistants/claude-code-proxy-on-aws/infra/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes --output-file infra/requirements.txt
alembic==1.18.4
    # via claude-code-proxy-on-aws
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via starlette
asyncpg==0.31.0
    # via claude-code-proxy-on-aws
attrs==25.4.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.263
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.1.1
    # via aws-cdk-lib
aws-cdk-cloud-assembly-schema==53.11.0
    # via aws-cdk-lib
aws-cdk-lib==2.246.0
    # via claude-code-proxy-on-aws
boto3==1.42.80
    # via claude-code-proxy-on-aws
botocore==1.42.80
    # via
    #   boto3
    #   s3transfer
cattrs==25.3.0
    # via jsii
colorama==0.4.6 ; sys_platform == 'win32'
    # via pytest
constructs==10.6.0
    # via
    #   aws-cdk-lib
    #   claude-code-proxy-on-aws
coverage==7.13.5
    # via pytest-cov
fastapi==0.135.2
    # via claude-code-proxy-on-aws
greenlet==3.3.2
    # via sqlalchemy
hypothesis==6.151.10
idna==3.11
    # via anyio
importlib-resources==6.5.2
    # via jsii
iniconfig==2.3.0
    # via pytest
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
jsii==1.127.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
packaging==26.0
    # via pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
pydantic==2.12.5
    # via
    #   claude-code-proxy-on-aws
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
pygments==2.20.0
    # via pytest
pytest==8.4.2
    # via
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.26.0
pytest-cov==6.3.0
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   jsii
ruff==0.11.13
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
sortedcontainers==2.4.0
    # via hypothesis
sqlalchemy==2.0.48
    # via
    #   alembic
    #   claude-code-proxy-on-aws
starlette==1.0.0
    # via fastapi
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.15.0
    # via
    #   alembic
    #   cattrs
    #   fastapi
    #   jsii
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via botocore
